requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiosignal==1.3.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.1.0
cryptography==43.0.1
datasets==2.21.0
dill==0.3.8
etelemetry==0.3.1
faiss-cpu==1.8.0.post1
filelock==3.16.0
fitz==0.0.1.dev2
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
huggingface-hub==0.24.6
idna==3.8
isodate==0.6.1
Jinja2==3.1.4
loadimg==0.3.1
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
prov==2.0.1
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydot==3.0.1
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.7.24
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.2.0
setuptools==74.1.2
simplejson==3.19.3
six==1.16.0
sympy==1.13.2
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
traits==6.3.2
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
xxhash==3.5.0
yarl==1.11.0
